% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-add_if_exists.R
\name{addIfExistLines}
\alias{addIfExistLines}
\title{Add Leaflet Polylines If They Exist}
\usage{
addIfExistLines(map = NULL, data, map_format = "pdf", ...)
}
\arguments{
\item{map}{a map widget object created from leaflet()}

\item{data}{the data object from which the argument values are derived.}

\item{...}{additional arguments passed to addPolylines()}
}
\description{
Add polylines to a leaflet map if there is sufficient data.
}
