% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-add_if_exists.R
\name{addIfExistMarkers}
\alias{addIfExistMarkers}
\title{Add Leaflet Markers If They Exist}
\usage{
addIfExistMarkers(map = NULL, data, map_format = "pdf", ...)
}
\arguments{
\item{map}{a map widget object created from leaflet()}

\item{data}{the data object from which the argument values are derived.}

\item{...}{additional arguments passed to leaflet::addCircleMarkers() or ggplot2::geom_sf()}
}
\description{
Add markers to a leaflet map if there is sufficient data.
}
