% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives_1D.r
\name{arc_node}
\alias{arc_node}
\alias{arc_node.sf}
\title{Arc node topology for simple features.}
\usage{
\method{arc_node}{sf}(x, ...)
}
\description{
Arc node topology for simple features.
}
\examples{
x <- sf::st_read(system.file("extdata/file.geojson", package= "sc"))
arc_node(x)  ## get the nodes
## now get the arcs (should the functions be called arc() and node()?)
}
