% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basins.R
\docType{data}
\name{basins}
\alias{basins}
\title{The eight major river basins of South Carolina}
\format{
A \code{sf and data.frame} with columns for basin name and geometry.
}
\usage{
basins
}
\description{
This data was acquired from
https://sc-department-of-health-and-environmental-control-gis-sc-dhec.hub.arcgis.com/datasets/sc-major-river-basins/explore?location=33.617600%2C-80.935600%2C8.34
in October 2021. The raw data was edited to extend a small coastal region of the Santee basin,
to match the latest SC boundary shapefile. Not all of the shapefile is edited to match the
coastal extent of the latest SC Boundary shapefile.
}
\keyword{datasets}
