% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-capacity_use.R
\docType{data}
\name{capacity_use}
\alias{capacity_use}
\title{capacity_use}
\format{
A \code{data.frame} with columns:
\describe{
\item{Name}{Name of capacity use zone}
\item{Date_Established}{date the capacity use zone was formally established}
\item{Counties}{A list column of counties in each zone}
\item{Aliases}{Other ways of spelling the Name of the zone}
}
}
\usage{
capacity_use
}
\description{
A table of capacity use zones in SC (not spatial)
}
\keyword{datasets}
