% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-capacity_use.R
\docType{data}
\name{capacity_use_counties}
\alias{capacity_use_counties}
\title{capacity_use_counties}
\format{
A \code{data.frame} with columns:
\describe{
\item{County}{Name of the county}
\item{cap_use_zone}{Name of capacity use zone}
}
}
\usage{
capacity_use_counties
}
\description{
A table to join counties and capacity use zones
}
\keyword{datasets}
