% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coordinates.R
\name{check_points_in_SC}
\alias{check_points_in_SC}
\title{Point in or outside the South Carolina state boundary.}
\usage{
check_points_in_SC(x)
}
\arguments{
\item{x}{is a point shapefile with two of its columns for longitude and}
}
\value{
a logical output (TRUE if point lies inside the polygon)
}
\description{
Checks if a point with longitude and latitude information, lies
}
