% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-obscure_coordinates.R
\name{obscure_coordinates}
\alias{obscure_coordinates}
\title{Obscure the latitude and longitude in a given shapefile}
\usage{
obscure_coordinates(x)
}
\arguments{
\item{x}{is a shapefile with 'sf' class}
}
\value{
a table of class sf and data.frame with geometry information,
}
\description{
The function obscures the original latitude and longitude
}
