% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-planning_basins.R
\docType{data}
\name{planning_basins}
\alias{planning_basins}
\title{The eight  planning basins in South Carolina}
\format{
A \code{sf and data.frame} with columns for basin name and geometry.
}
\usage{
planning_basins
}
\description{
This data contains a shapefile for eight planning basins in South Carolina, namely "Broad,"
"Catawba," "Edisto," "Pee Dee," "Upper Savannah," "Saluda," "Lower Savannah-Salkehatchie," and "Santee."
}
\keyword{datasets}
