% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-update_bbox.R
\name{replace_point_coords}
\alias{replace_point_coords}
\title{Replace Selected Point Coordinates}
\usage{
replace_point_coords(sf, index, x = NA, y = NA)
}
\arguments{
\item{sf}{an object with class 'sf' or 'sfc'}

\item{x}{the new x coordinate (longitude), or NA to leave as is.}

\item{y}{the new y coordinate (latitude), or NA to leave as is.}
}
\value{
sf, but with updated point coordinates
}
\description{
Replace the point geometry for selected index in an sf object.
}
