% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_basins.R
\docType{data}
\name{SC_basins}
\alias{SC_basins}
\title{South Carolina state's eight major basins shapefile.}
\format{
A \code{sf and data.frame} with geometry and multiple column
\describe{
\item{Basin8}{Names of the eight basins}
\item{}{}
}
}
\usage{
SC_basins
}
\description{
The shapefile will be used for checking location errors for
withdrawal points of the sourceids and fill in missing basin information for
sourceids with coordinate information.
}
\keyword{datasets}
