% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_boundary_new.R
\docType{data}
\name{SC_boundary_new}
\alias{SC_boundary_new}
\title{South Carolina state boundary shapefile documentation}
\format{
A \code{sf and data.frame} with geometry and multiple column
\describe{
\item{XCoord}{longitude}
\item{YCoord}{latitude}
}
}
\usage{
SC_boundary_new
}
\description{
The shapefile will be used for checking location errors for
withdrawal points of the sourceids. This .R file is created to document the dataset
and is similar to documenting a function, but instead here the name of the dataset
is used to document it. The shapefile also includes the 3 mile coastal extension area.
}
\keyword{datasets}
