% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives_1D.r
\name{sc_node}
\alias{sc_node}
\alias{sc_node.sf}
\title{Arc node topology for simple features.}
\usage{
\method{sc_node}{sf}(x, ...)
}
\arguments{
\item{x}{input simple features object}

\item{...}{arguments for methods}
}
\description{
Arc node topology for simple features.
}
\examples{
# x <- sf::st_read(system.file("extdata/file.geojson", package= "sc"))
#sc_node(x)  ## get the nodes
## now get the arcs (should the functions be called arc() and node()?)
}
