% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-forms.r
\name{sf}
\alias{sf}
\alias{sf}
\alias{sf.PRIMITIVE}
\alias{sf}
\alias{sf.PATH}
\title{Recompose `sf` simple features from `PRIMITIVE`` models.}
\usage{
sf(x)

\method{sf}{PRIMITIVE}(x, ...)

\method{sf}{PATH}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}
}
\value{
`PRIMITIVE`
}
\description{
Recompose `sf` simple features from `PRIMITIVE`` models.
}
\examples{
prim <- PRIMITIVE(inlandwaters)
library(sf)
\donttest{
\dontrun{
plot(sf(prim))
}
}
}
