% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-update_bbox.R
\name{st_bbox_update}
\alias{st_bbox_update}
\title{Update Bounding Box}
\usage{
st_bbox_update(x)
}
\arguments{
\item{x}{an object with class 'sf' or 'sfc'}
}
\value{
x, but with an updated bounding box
}
\description{
Update the bounding box of an 'sf' or 'sfc' object. Bounding box information
is typically handled internally by the sf package. However, if you modify a geometry directly,
the bounding box is not necessarily updated to reflect the modifications. This function can
be used to provide an updated bounding box.
}
