% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-drop_geometry.R
\name{st_drop_geometry_if}
\alias{st_drop_geometry_if}
\title{Drop geometry}
\usage{
st_drop_geometry_if(x)
}
\arguments{
\item{x}{a shape file with class 'sf' and 'data.frame'}
}
\value{
a data.frame
}
\description{
the function can remove geometry from a data set with class 'sf'
 i.e removes the crs and geometry information.
'SC_boundary' is a shape file with class 'sf' and 'data.frame'.
When the shape file is viewed it has two polygons, each with attributes
(columns of) for 'State', 'Area', 'Name', 'Shape_Area', 'Shape_Len', 'XCoord',
'YCoord', 'geometry', and CRS information.
After using this function, the attribute columns are preserved, except
geometry column. Thereby converting the shape file into a regular data frame.
}
