% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-drop_geometry.R
\name{st_drop_geometry_if}
\alias{st_drop_geometry_if}
\title{Drop geometry}
\usage{
st_drop_geometry_if(x)
}
\arguments{
\item{x}{a shape file with class 'sf' and 'data.frame'}
}
\value{
a data.frame
}
\description{
the function can remove geometry from a data set with class 'sf'
}
